refactor: extract connection managers outside of Db
