+++ packages/jazz-tools/src/runtime/connection-manager/index.ts
@@ -1,0 +1,7 @@
+export type {
+  ConnectionManager as ConnectionBridge,
+  ConnectionBridgeClientInput,
+  ConnectionManagerHost as ConnectionBridgeHost,
+} from "./types.js";
+export { DirectConnectionManager as DirectConnectionBridge } from "./direct-connection-manager.js";
+export { BrowserConnectionManager as BrowserBrokerConnectionBridge } from "./browser-connection-manager.js";
